fixed RpcAgentAvailabilityServiceRemote.java
