+++ admin-service-api/src/main/java/hu/schonherz/project/admin/service/api/rpc/RpcAgentAvailabilityServiceRemote.java
@@ -2,4 +2,6 @@ package hu.schonherz.project.admin.service.api.rpc;
 
 public interface RpcAgentAvailabilityServiceRemote {
     Long getAvailableAgent(String source) throws NoAvailableAgentFoundException;
-}
+    
+    void SetAgentAvailability(String username) throws NoAvailableAgentFoundException;
+    }
