--- admin-service-api/src/main/java/hu/schonherz/project/admin/service/api/rpc/RpcSetAgentAvailabilityServiceRemote.java
@@ -1,5 +1,0 @@
-package hu.schonherz.project.admin.service.api.rpc;
-
-public interface RpcSetAgentAvailabilityServiceRemote {
-    void SetAgentAvailability(String username) throws NoAvailableAgentFoundException;
-}
