+++ admin-service/src/main/java/hu/schonherz/project/admin/service/impl/rpc/RpcAgentAvailabilityServiceBean.java
@@ -5,18 +5,25 @@ import javax.ejb.EJB;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import hu.schonherz.project.admin.data.repository.CompanyRepository;
+import hu.schonherz.project.admin.data.repository.UserRepository;
 import hu.schonherz.project.admin.service.api.rpc.NoAvailableAgentFoundException;
 import hu.schonherz.project.admin.service.api.rpc.RpcAgentAvailabilityServiceRemote;
 import hu.schonherz.project.admin.service.api.service.company.CompanyServiceLocal;
+import hu.schonherz.project.admin.service.api.service.user.UserServiceLocal;
 import hu.schonherz.project.admin.service.api.vo.CompanyVo;
 import hu.schonherz.project.admin.service.api.vo.UserVo;
 import hu.schonherz.project.admin.service.mapper.company.CompanyEntityVoMapper;
+import hu.schonherz.project.admin.service.mapper.user.UserEntityVoMapper;
 
 public class RpcAgentAvailabilityServiceBean implements RpcAgentAvailabilityServiceRemote {
     @Autowired
     private CompanyRepository companyRepository;
     @EJB
     private CompanyServiceLocal companyServiceLocal;
+    @Autowired
+    private UserRepository userRepository;
+    @EJB
+    private UserServiceLocal userServiceLocal;
 
     @Override
     public Long getAvailableAgent(final String source) throws NoAvailableAgentFoundException {
@@ -28,5 +35,12 @@ public class RpcAgentAvailabilityServiceBean implements RpcAgentAvailabilityServ
             }
         }
         throw new NoAvailableAgentFoundException("There's no available Agent with " + source + " domain.");
+    }
+
+    @Override
+    public void SetAgentAvailability(final String username) throws NoAvailableAgentFoundException {
+        UserVo userVo = UserEntityVoMapper.toVo(userRepository.findByUsername(username));
+        userVo.setAvailable(true);
+        userRepository.save(UserEntityVoMapper.toEntity(userVo));
     }
 }
